requirements.txt
tqdm
Pillow
hydra-core
torch
torchvision
kornia